adds more tests

+++ v1/lexer/lexer_test.go
@@ -13,27 +13,27 @@ func TestLexer(t *testing.T) {
 		tokens []core.Token
 	}{
 		{name: "empty-string", input: "     ", tokens: nil},
-		{name: "single-digit", input: "1", tokens: []core.Token{{Kind: core.INT, Value: "1"}}},
-		{name: "triple-digit", input: "123   ", tokens: []core.Token{{Kind: core.INT, Value: "123"}}},
+		{name: "single-digit", input: "1", tokens: []core.Token{{Kind: core.FLOAT, Value: "1"}}},
+		{name: "triple-digit", input: "123   ", tokens: []core.Token{{Kind: core.FLOAT, Value: "123"}}},
 		{name: "two-triple-digits", input: "123   567", tokens: []core.Token{
-			{Kind: core.INT, Value: "123"},
-			{Kind: core.INT, Value: "567"},
+			{Kind: core.FLOAT, Value: "123"},
+			{Kind: core.FLOAT, Value: "567"},
 		}},
 		{name: "add-two-ints", input: "789 +47", tokens: []core.Token{
-			{Kind: core.INT, Value: "789"},
+			{Kind: core.FLOAT, Value: "789"},
 			{Kind: core.ADD, Value: "+"},
-			{Kind: core.INT, Value: "47"},
+			{Kind: core.FLOAT, Value: "47"},
 		}},
 		{name: "add-and-sub-some-ints", input: "7 - 1+22-334+99", tokens: []core.Token{
-			{Kind: core.INT, Value: "7"},
+			{Kind: core.FLOAT, Value: "7"},
 			{Kind: core.SUB, Value: "-"},
-			{Kind: core.INT, Value: "1"},
+			{Kind: core.FLOAT, Value: "1"},
 			{Kind: core.ADD, Value: "+"},
-			{Kind: core.INT, Value: "22"},
+			{Kind: core.FLOAT, Value: "22"},
 			{Kind: core.SUB, Value: "-"},
-			{Kind: core.INT, Value: "334"},
+			{Kind: core.FLOAT, Value: "334"},
 			{Kind: core.ADD, Value: "+"},
-			{Kind: core.INT, Value: "99"},
+			{Kind: core.FLOAT, Value: "99"},
 		}},
 	}
